Added methods for adding/removing roles from users

+++ src/users/users.service.ts
@@ -1,10 +1,15 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { ModelClass } from 'objection';
+import { ModelClass, transaction } from 'objection';
 import { UserModel } from 'src/database/models/user.model';
+import { UsersRolesService } from './user-roles.service';
 
 @Injectable()
 export class UsersService {
-  constructor(@Inject('UserModel') private modelClass: ModelClass<UserModel>) {}
+  constructor(
+    @Inject('UserModel') private modelClass: ModelClass<UserModel>,
+    private userRolesService: UsersRolesService,
+  ) {}
+
   findAll() {
     return this.modelClass.query().select('id', 'email');
   }
@@ -34,12 +39,26 @@ export class UsersService {
       .first();
   }
 
-  delete(id: number) {
-    return this.modelClass
-      .query()
-      .delete()
-      .where({ id })
-      .returning('*')
-      .first();
+  async delete(id: number) {
+    return transaction(this.modelClass, async (_, tx) => {
+      await this.userRolesService.deleteByUserId(id).transacting(tx);
+      return this.modelClass
+        .query()
+        .delete()
+        .where({ id })
+        .returning('*')
+        .first()
+        .transacting(tx);
+    });
+  }
+
+  async addRole(userId: number, roleId: number) {
+    await this.userRolesService.create({ userId, roleId });
+    return this.findOne(userId);
+  }
+
+  async removeRole(userId: number, roleId: number) {
+    await this.userRolesService.delete({ userId, roleId });
+    return this.findOne(userId);
   }
 }
